requirements.txt
pika==1.2.0
python-json-logger==2.0.2
PyYAML==6.0
structlog==21.5.0
viaa-chassis==0.2.0rc1
bagit==1.8.1
pulsar-client==2.8.1
meemoo-cloudevents==0.1.0rc3
# Dev
black==21.12b0
flake8==4.0.1
# Test
pytest==6.2.5
pytest-cov==3.0.0
